put method inside utils
