+++ antarest/study/business/output/aggregator_management.py
@@ -18,10 +18,7 @@ from typing import Any, Dict, Iterator, List, MutableSequence, Optional, Sequenc
 import pandas as pd
 
 from antarest.core.exceptions import MCRootNotHandled, OutputAggregationError, OutputNotFound, OutputSubFolderNotFound
-from antarest.study.storage.rawstudy.model.filesystem.matrix.date_serializer import (
-    FactoryDateSerializer,
-    rename_unnamed,
-)
+from antarest.study.business.output.utils import parse_output_file
 from antarest.study.storage.rawstudy.model.filesystem.matrix.matrix import MatrixFrequency
 
 # noinspection SpellCheckingInspection
@@ -146,18 +143,8 @@ class AggregatorManager:
             else MCRoot.MC_ALL
         )
 
-    @staticmethod
-    def parse_output_file(file_path: Path, frequency: MatrixFrequency, n_rows: int | None = None) -> pd.DataFrame:
-        csv_file = pd.read_csv(
-            file_path, sep="\t", skiprows=4, header=[0, 1, 2], na_values="N/A", float_precision="legacy", nrows=n_rows
-        )
-        date_serializer = FactoryDateSerializer.create(frequency.value, "")
-        _, body = date_serializer.extract_date(csv_file)
-        rename_unnamed(body)
-        return body
-
     def _parse_output_file(self, file_path: Path, normalize_column_name: bool = True) -> pd.DataFrame:
-        body = self.parse_output_file(file_path, self.frequency)
+        body = parse_output_file(file_path, self.frequency)
 
         df = body.astype(float)
